+++ tests/conftest.py
@@ -102,7 +102,7 @@ def fixture_analysis_family_single(case_id: str, family_name: str, ticket_nr: in
     return family
 
 
-@pytest.yield_fixture(scope="function", name="analysis_family")
+@pytest.fixture(scope="function", name="analysis_family")
 def fixture_analysis_family(case_id: str, family_name: str, ticket_nr: int) -> dict:
     """Return a dictionary with information from a analysis family"""
     family = {
@@ -201,13 +201,13 @@ def fixture_genotype_api(genotype_config: dict) -> GenotypeAPI:
     return _genotype_api
 
 
-@pytest.yield_fixture(scope="function")
+@pytest.fixture(scope="function")
 def madeline_api(madeline_output) -> MockMadelineAPI:
     """madeline_api fixture"""
     _api = MockMadelineAPI()
     _api.set_outpath(madeline_output)
 
-    yield _api
+    return _api
 
 
 # Files fixtures
@@ -614,14 +614,13 @@ def fixture_compress_hk_fastq_bundle(
     return hk_bundle_data
 
 
-@pytest.yield_fixture(scope="function", name="housekeeper_api")
 def fixture_housekeeper_api(hk_config_dict: dict) -> MockHousekeeperAPI:
     """Setup Housekeeper store."""
     _api = MockHousekeeperAPI(hk_config_dict)
     yield _api
 
 
-@pytest.yield_fixture(scope="function", name="real_housekeeper_api")
+@pytest.fixture(scope="function", name="real_housekeeper_api")
 def fixture_real_housekeeper_api(hk_config_dict: dict) -> HousekeeperAPI:
     """Setup a real Housekeeper store."""
     _api = HousekeeperAPI(hk_config_dict)
@@ -629,7 +628,7 @@ def fixture_real_housekeeper_api(hk_config_dict: dict) -> HousekeeperAPI:
     yield _api
 
 
-@pytest.yield_fixture(scope="function", name="populated_housekeeper_api")
+@pytest.fixture(scope="function", name="populated_housekeeper_api")
 def fixture_populated_housekeeper_api(
     housekeeper_api: MockHousekeeperAPI, hk_bundle_data: dict, helpers
 ) -> MockHousekeeperAPI:
@@ -639,7 +638,7 @@ def fixture_populated_housekeeper_api(
     return hk_api
 
 
-@pytest.yield_fixture(scope="function", name="hk_version_obj")
+@pytest.fixture(scope="function", name="hk_version_obj")
 def fixture_hk_version_obj(
     housekeeper_api: MockHousekeeperAPI, hk_bundle_data: dict, helpers
 ) -> MockHousekeeperAPI:
@@ -678,7 +677,7 @@ def fixture_hermes_api(hermes_process: ProcessMock) -> HermesApi:
 # Scout fixtures
 
 
-@pytest.yield_fixture(scope="function", name="scout_api")
+@pytest.fixture(scope="function", name="scout_api")
 def fixture_scout_api() -> MockScoutAPI:
     """Setup Scout api."""
     _api = MockScoutAPI()
@@ -688,7 +687,7 @@ def fixture_scout_api() -> MockScoutAPI:
 # Crunchy fixtures
 
 
-@pytest.yield_fixture(scope="function", name="crunchy_api")
+@pytest.fixture(scope="function", name="crunchy_api")
 def fixture_crunchy_api():
     """Setup Crunchy api."""
     _api = MockCrunchyAPI()
@@ -698,7 +697,7 @@ def fixture_crunchy_api():
 # Store fixtures
 
 
-@pytest.yield_fixture(scope="function", name="analysis_store")
+@pytest.fixture(scope="function", name="analysis_store")
 def fixture_analysis_store(
     base_store: Store, analysis_family: dict, wgs_application_tag: str, helpers
 ):
@@ -710,14 +709,14 @@ def fixture_analysis_store(
     yield base_store
 
 
-@pytest.yield_fixture(scope="function", name="analysis_store_trio")
+@pytest.fixture(scope="function", name="analysis_store_trio")
 def fixture_analysis_store_trio(analysis_store):
     """Setup a store instance with a trion loaded for testing analysis API."""
 
     yield analysis_store
 
 
-@pytest.yield_fixture(scope="function", name="analysis_store_single_case")
+@pytest.fixture(scope="function", name="analysis_store_single_case")
 def fixture_analysis_store_single(base_store, analysis_family_single_case, helpers):
     """Setup a store instance with a single ind case for testing analysis API."""
     helpers.ensure_family_from_dict(base_store, family_info=analysis_family_single_case)
@@ -725,13 +724,13 @@ def fixture_analysis_store_single(base_store, analysis_family_single_case, helpe
     yield base_store
 
 
-@pytest.yield_fixture(scope="function", name="customer_group")
+@pytest.fixture(scope="function", name="customer_group")
 def fixture_customer_group() -> str:
     """Return a default customer group"""
     return "all_customers"
 
 
-@pytest.yield_fixture(scope="function", name="customer_production")
+@pytest.fixture(scope="function", name="customer_production")
 def fixture_customer_production(customer_group) -> dict:
     """Return a dictionary with information about the prod customer"""
     _cust = dict(
@@ -743,13 +742,13 @@ def fixture_customer_production(customer_group) -> dict:
     return _cust
 
 
-@pytest.yield_fixture(scope="function", name="external_wgs_application_tag")
+@pytest.fixture(scope="function", name="external_wgs_application_tag")
 def fixture_external_wgs_application_tag() -> str:
     """Return the external wgs app tag"""
     return "WGXCUSC000"
 
 
-@pytest.yield_fixture(scope="function", name="external_wgs_info")
+@pytest.fixture(scope="function", name="external_wgs_info")
 def fixture_external_wgs_info(external_wgs_application_tag) -> dict:
     """Return a dictionary with information external WGS application"""
     _info = dict(
@@ -762,13 +761,13 @@ def fixture_external_wgs_info(external_wgs_application_tag) -> dict:
     return _info
 
 
-@pytest.yield_fixture(scope="function", name="external_wes_application_tag")
+@pytest.fixture(scope="function", name="external_wes_application_tag")
 def fixture_external_wes_application_tag() -> str:
     """Return the external whole exome sequencing app tag"""
     return "EXXCUSR000"
 
 
-@pytest.yield_fixture(scope="function", name="external_wes_info")
+@pytest.fixture(scope="function", name="external_wes_info")
 def fixture_external_wes_info(external_wes_application_tag) -> dict:
     """Return a dictionary with information external WES application"""
     _info = dict(
@@ -781,13 +780,13 @@ def fixture_external_wes_info(external_wes_application_tag) -> dict:
     return _info
 
 
-@pytest.yield_fixture(scope="function", name="wgs_application_tag")
+@pytest.fixture(scope="function", name="wgs_application_tag")
 def fixture_wgs_application_tag() -> str:
     """Return the wgs app tag"""
     return "WGSPCFC060"
 
 
-@pytest.yield_fixture(scope="function", name="wgs_application_info")
+@pytest.fixture(scope="function", name="wgs_application_info")
 def fixture_wgs_application_info(wgs_application_tag) -> dict:
     """Return a dictionary with information the WGS application"""
     _info = dict(
@@ -802,7 +801,7 @@ def fixture_wgs_application_info(wgs_application_tag) -> dict:
     return _info
 
 
-@pytest.yield_fixture(scope="function", name="store")
+@pytest.fixture(scope="function", name="store")
 def fixture_store() -> Store:
     """Fixture with a CG store"""
     _store = Store(uri="sqlite:///")
@@ -816,7 +815,7 @@ def fixture_apptag_rna() -> str:
     return "RNAPOAR025"
 
 
-@pytest.yield_fixture(scope="function", name="base_store")
+@pytest.fixture(scope="function", name="base_store")
 def fixture_base_store(store: Store, apptag_rna: str) -> Store:
     """Setup and example store."""
     customer_group = store.add_customer_group("all_customers", "all customers")
@@ -1009,7 +1008,7 @@ def sample_store(base_store) -> Store:
     return base_store
 
 
-@pytest.yield_fixture(scope="function")
+@pytest.fixture(scope="function")
 def disk_store(cli_runner, invoke_cli) -> Store:
     """Store on disk"""
     database = "./test_db.sqlite3"